feat(checkpoint): reorganize checkpoint code and improve robustness
