+++ example/gpt2/config.h
@@ -2,7 +2,7 @@
 
 #include "infini_train/include/nn/modules/transformer/transformer_config.h"
 
-namespace nn = infini_train::nn;
+namespace infini_train {
 namespace gpt2 {
 inline nn::TransformerConfig GPT2Config() {
     return {.block_size = 1024,
@@ -22,5 +22,5 @@ inline nn::TransformerConfig GPT2Config() {
             .ffn_dim_multiplier = std::nullopt,
             .multiple_of = 1};
 }
-
 } // namespace gpt2
+} // namespace infini_train
